requirements.txt
GitPython
streamlit
pandas
matplotlib